% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHighlightedDF.R
\name{GetHighlightedDF}
\alias{GetHighlightedDF}
\title{GetHighlightedDF [Credit to: https://github.com/holstius]}
\usage{
GetHighlightedDF()
}
\value{
Highlighted data frame
}
\description{
GetHighlightedDF [Credit to: https://github.com/holstius]
}

