% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrLeft.R
\name{InsertPipeDplyrLeft}
\alias{InsertPipeDplyrLeft}
\title{Insert \%<\%.}
\usage{
InsertPipeDplyrLeft()
}
\description{
Call this function as an addin to insert \code{ \%<\% } at the cursor position.
}
