% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrNot.R
\name{InsertPipeDplyrNot}
\alias{InsertPipeDplyrNot}
\title{Insert \%!>\%.}
\usage{
InsertPipeDplyrNot()
}
\description{
Call this function as an addin to insert \code{ \%!>\% } at the cursor position.
}
