% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrPlusLeft.R
\name{InsertPipeDplyrPlusLeft}
\alias{InsertPipeDplyrPlusLeft}
\title{Insert \%<+\%.}
\usage{
InsertPipeDplyrPlusLeft()
}
\description{
Call this function as an addin to insert \code{ \%<+\% } at the cursor position.
}
