% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrPlusRight.R
\name{InsertPipeDplyrPlusRight}
\alias{InsertPipeDplyrPlusRight}
\title{Insert \%+>\%.}
\usage{
InsertPipeDplyrPlusRight()
}
\description{
Call this function as an addin to insert \code{ \%+>\% } at the cursor position.
}
