% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrTee.R
\name{InsertPipeDplyrTee}
\alias{InsertPipeDplyrTee}
\title{Insert \%T>\%.}
\usage{
InsertPipeDplyrTee()
}
\description{
Call this function as an addin to insert \code{ \%T>\% } at the cursor position.
}
