% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeDplyrVar.R
\name{InsertPipeDplyrVar}
\alias{InsertPipeDplyrVar}
\title{Insert \%$\%.}
\usage{
InsertPipeDplyrVar()
}
\description{
Call this function as an addin to insert \code{ \%$\% } at the cursor position.
}
