% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickGlimpseDataFrame.R
\name{QuickGlimpseDataFrame}
\alias{QuickGlimpseDataFrame}
\title{Quick Glimpse Data Frame}
\usage{
QuickGlimpseDataFrame(h = GetHighlightedDF())
}
\arguments{
\item{df}{Data frame for rapid glimpsing}
}
\value{
NA
}
\description{
Quick Glimpse Data Frame
}
