% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewDataFrame.R
\name{ViewDataFrame}
\alias{ViewDataFrame}
\title{View Data Frame}
\usage{
ViewDataFrame(df = GetHighlightedDF())
}
\arguments{
\item{df}{Data frame for rapid viewing}
}
\value{
NA
}
\description{
View Data Frame
}

